+++ src/main/java/net/sf/jabref/logic/undo/UndoRedoEvent.java
@@ -4,52 +4,10 @@ package net.sf.jabref.logic.undo;
  * Event sent when something is undone or redone
  *
  */
-public class UndoRedoEvent {
-
-    private final boolean canUndo;
-    private final String undoDescription;
-    private final boolean canRedo;
-    private final String redoDescription;
-
+public class UndoRedoEvent extends UndoChangeEvent {
 
     public UndoRedoEvent(boolean canUndo, String undoDescription, boolean canRedo, String redoDescription) {
-        this.canUndo = canUndo;
-        this.undoDescription = undoDescription;
-        this.canRedo = canRedo;
-        this.redoDescription = redoDescription;
-    }
-
-    /**
-     *
-     * @return true if there is an action that can be undone
-     */
-    public boolean isCanUndo() {
-        return canUndo;
-    }
-
-    /**
-     *
-     * @return A description of the action to be undone
-     */
-    public String getUndoDescription() {
-        return undoDescription;
-    }
-
-    /**
-     *
-     * @return true if there is an action that can be redone
-     */
-    public boolean isCanRedo() {
-        return canRedo;
-    }
-
-
-    /**
-    *
-    * @return A description of the action to be redone
-    */
-    public String getRedoDescription() {
-        return redoDescription;
+        super(canUndo, undoDescription, canRedo, redoDescription);
     }
 
 }
